libs: bufferproxy: add gst_vaapi_buffer_proxy_{set,get}_mem()
This patch adds an GstMemory as a variable member of the buffer proxy,
because we will need to associate the buffer proxy with the memory
which exposes it. Later, we will know which memory, in the video buffer
pool, is attached to the processed surface.

Also, its pair of getter and setter API calls.

https://bugzilla.gnome.org/show_bug.cgi?id=755072