gstvaapivideobufferpool: set correct buffer size when configuring the pool
Otherwise the buffer is discarded when returning to the pool.
Indeed when ref count of a GstBuffer reaches 0, then the refered
pool will queue it (and ref it) only if the buffer size matches
the configured buffer size on the pool.
Can be debug with GST_DEBUG=*PERF*:6, see gstbufferpool.c

https://bugzilla.gnome.org/show_bug.cgi?id=755072