vaapivideomemory: call gst_vaapi_buffer_proxy_release_data if GST_PAD_SRC
The surface is an output so it is filled by a gst vaapi element. So the
va driver needs write access on the surface. The following call will
released the derived image so mark the surface as unbusy. This is how it
has to be done as discussed on libva mailing list.

https://bugzilla.gnome.org/show_bug.cgi?id=755072