vaapivideomemory: release proxy's data if downstream
The created downstream surface is going to be filled by VAAPI elements.
So, the driver needs write access on that surface.

This patch releases the holded derived image in the proxy, so the
surface will be marked as unbusy.

This is how it has to be done as discussed on libva mailing list.

https://bugzilla.gnome.org/show_bug.cgi?id=755072